+++ ZEA.Applications.Workflows/IPreProcessBehavior.cs
@@ -1,0 +1,11 @@
+using ZEA.Techniques.ADTs;
+using ZEA.Techniques.ADTs.Helpers;
+
+namespace ZEA.Applications.Workflows;
+
+public interface IPreProcessBehavior<in TRequest, TError>
+{
+	Task<Either<Success, TError>> ExecuteAsync(
+		TRequest request,
+		CancellationToken cancellationToken);
+}